+++ composer.json
@@ -26,6 +26,7 @@
         "sentry/sentry-laravel": "^2.13",
         "silber/bouncer": "^1.0",
         "spatie/data-transfer-object": "^3.7",
+        "spatie/laravel-data": "^2.0",
         "symfony/http-client": "^6.1",
         "symfony/mailgun-mailer": "^6.1",
         "vladimir-yuldashev/laravel-queue-rabbitmq": "^12.0",
