+++ src/Kanvas/Filesystem/Models/FilesystemSettings.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Filesystem\Filesystem\Models;
+namespace Kanvas\Filesystem\Models;
 
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Kanvas\Models\BaseModel;
 
 /**
@@ -33,6 +34,6 @@ class FilesystemSettings extends BaseModel
      */
     public function filesystem() : BelongsTo
     {
-        return $this->belongsTo(Filesystem::class, 'filesystem_id');
+        return $this->belongsTo(FileSystem::class, 'filesystem_id');
     }
 }
